Extract MCP core crate
