+++ codex-rs/Cargo.toml
@@ -32,6 +32,7 @@ members = [
     "linux-sandbox",
     "lmstudio",
     "login",
+    "mcp-core",
     "mcp-server",
     "network-proxy",
     "ollama",
@@ -114,6 +115,7 @@ codex-keyring-store = { path = "keyring-store" }
 codex-linux-sandbox = { path = "linux-sandbox" }
 codex-lmstudio = { path = "lmstudio" }
 codex-login = { path = "login" }
+codex-mcp = { path = "mcp-core" }
 codex-mcp-server = { path = "mcp-server" }
 codex-network-proxy = { path = "network-proxy" }
 codex-ollama = { path = "ollama" }
