+++ codex-rs/mcp-core/BUILD.bazel
@@ -1,0 +1,9 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "mcp-core",
+    crate_name = "codex_mcp",
+    compile_data = [
+        "src/consequential_tool_message_templates.json",
+    ],
+)
