+++ codex-rs/mcp-core/src/connection_manager_tests.rs
