+++ codex-rs/mcp-core/src/consequential_tool_message_templates.json
